% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyweb.R
\name{easyweb}
\alias{easyweb}
\title{Create an easyweb rmarkdown website}
\usage{
easyweb(web_tsv, path = tempdir(), publish = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{web_tsv}{The path to a tab-separated file with `tag`s on the first
column and `value`s on the second. Check the example table in the package
for a template.}

\item{path}{The path where you website files will be saved to.}

\item{publish}{Whether to publish your website (currently not implemented).}

\item{overwrite}{Whether to overwrite the files.}
}
\value{
The path to your website.
}
\description{
This function takes as input a two column table with a set of `tag`s and
`value`s that are used to construct a rmarkdown website that you can then
modify yourself and publish. This website will help you get your online
presence started!
}
\examples{

## Example easyweb table
options(width = 200)
easyweb_example

## Create an example website directory where the files will be saved
example_website_path <- file.path(tempdir(), 'easyweb')
dir.create(example_website_path, showWarnings = FALSE)
example_tsv <- file.path(example_website_path, 'web.tsv')

## Save the example table as a file
write.table(easyweb_example, file = example_tsv, sep = '\\t', quote = FALSE, row.names = FALSE)

## Build the website
easyweb(example_tsv, path = example_website_path, overwrite = TRUE)

}
